move require commands stuff to requiredcmds.go
